code cleanup & enable trace ("-t") for MBR decoding ("-mbr")
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder@3246 1f5c12ca-751b-0410-a591-d2e778427230

+++ trunk/moses/src/WordsBitmap.h
@@ -58,7 +58,7 @@ protected:
 	//sets elements by vector
 	void Initialize(std::vector<bool> vector)
 	{
-		int vector_size = vector.size();
+		size_t vector_size = vector.size();
 		for (size_t pos = 0 ; pos < m_size ; pos++)
 		{
 			if (pos < vector_size && vector[pos] == true) m_bitmap[pos] = true;
